Finale

+++ CarDealer.API/Services/CustomerService.cs
@@ -3,151 +3,188 @@ using CarDealer.Data;
 using CarDealer.Data.Entities;
 using CarDealer.Services.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using System.Diagnostics.Metrics;
+using System.Net;
 
 namespace CarDealer.Services
 {
     public class CustomerService : ICustomerService
     {
+        #region Setup
         private readonly CarDealerContext _context;
         private readonly ILogger<CustomerService> _logger;
-
 
         public CustomerService(CarDealerContext context, ILogger<CustomerService> logger)
         {
             _context = context;
             _logger = logger;
         }
-   
+        #endregion
 
-        public async Task<Customer> GetByID(Guid customerID)
+
+        #region Get By ID
+        public async Task<Customer> GetByID(Guid guid)
         {
             try
             {
-                Customer customer = await _context.Customers.FindAsync(customerID);
-
-                if (customer == null)
+                if (ExistsByID(guid))
                 {
-                    _logger.LogWarning($"Customer: {customerID}, not found.");
+                    Customer? customer = await _context.Customers.FindAsync(guid);
+                    _logger.LogInformation($"Customer: {guid}, has been retrieved successfully.");
+                    return customer;
                 }
                 else
                 {
-                    _logger.LogInformation($"Customer: {customerID}, has been retrieved successfully.");
+                    _logger.LogWarning($"Sale: {guid}, not found.");
+                    return null;
                 }
-
-                return customer;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"An error occurred while getting the Customer with ID: {customerID}.");
+                _logger.LogError(ex, $"An error occurred while getting sale: {guid}.");
                 return null;
             }
         }
+        #endregion
 
+       
+        #region Get By TIN
         public async Task<Customer> GetByTIN(string tin)
         {
             try
             {
-                Customer customer = await _context.Customers.FindAsync(tin);
-
-                if (customer == null)
+                if (ExistsTIN(tin))
                 {
-                    _logger.LogWarning($"Customer: {tin}, not found.");
+                    Customer? customer = await _context.Customers
+                    .FirstOrDefaultAsync(c => c.TIN == tin);
+
+                    _logger.LogInformation($"Customer: {tin}, has been retrieved successfully.");
+                    return customer;
                 }
                 else
                 {
-                    _logger.LogInformation($"Customer: {tin}, has been retrieved successfully.");
+                    _logger.LogWarning($"Customer: {tin}, not found.");
+                    return null;
                 }
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"An error occurred while getting car with plate: {tin}.");
+                return null;
+            }
+        }
+        #endregion
 
+
+        #region Get By First Name
+        public async Task<List<Customer>> GetByFirstName(string name)
+        {
+            try
+            {
+                List<Customer> customersList = await _context.Customers
+                .Where(a => a.First_Name == name)
+                .ToListAsync();
+
+                _logger.LogInformation($"Here is the Customer List named {name}");
+                return customersList;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"An error occurred while getting Customers named {name}.");
+                return null;
+            }
+        }
+        #endregion
+
+
+        #region Get By Middle Name
+        public async Task<List<Customer>> GetByMiddleName(string name)
+        {
+            try
+            {
+                List<Customer> customersList = await _context.Customers
+                .Where(a => a.Middle_Name == name)
+                .ToListAsync();
+
+                _logger.LogInformation($"Here is the Customer List named {name}");
+                return customersList;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"An error occurred while getting Customers named {name}.");
+                return null;
+            }
+        }
+        #endregion
+
+
+        #region Get By Last Name
+        public async Task<List<Customer>> GetByLastName(string name)
+        {
+            try
+            {
+                List<Customer> customersList = await _context.Customers
+                .Where(a => a.Last_Name == name)
+                .ToListAsync();
+
+                _logger.LogInformation($"Here is the Customer List named {name}");
+                return customersList;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"An error occurred while getting Customers named {name}.");
+                return null;
+            }
+        }
+        #endregion
+
+
+        #region Add Customer
+        public async Task<Customer> AddCustomer(CustomerData customerData)
+        {
+            try
+            {
+                Customer customer = new Customer
+                {
+                    TIN = customerData.TIN,
+                    First_Name = customerData.First_Name,
+                    Middle_Name = customerData.Middle_Name,
+                    Last_Name = customerData.Last_Name,
+                    Date_Of_Birth = customerData.Date_Of_Birth,
+                    Phone_Number = customerData.Phone_Number,
+                    Email = customerData.Email,
+                    Address = customerData.Address,
+                    City = customerData.City,
+                    State = customerData.State,
+                    Zip_Code = customerData.Zip_Code,
+                    Country = customerData.Country,
+                    Registration_Date = customerData.Registration_Date
+                };
+
+                _context.Customers.Add(customer);
+                await _context.SaveChangesAsync();
+
+                _logger.LogInformation($"Customer added successfully: {customer}");
                 return customer;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"An error occurred while getting the Customer with ID: {tin}.");
+                _logger.LogError(ex, $"Error occurred while adding the Customer: {customerData}");
                 return null;
             }
         }
+        #endregion
 
-        public async Task<List<Customer>> GetByFirstName(string firstName)
+
+        #region Edit By Customer ID
+        public async Task<Customer> EditByID(Guid guid, CustomerData customerData)
         {
             try
             {
-                _logger.LogInformation($"The Customers named: {firstName}, have been found.");
-                return await _context.Customers
-                    .Where(a => a.First_Name.ToLower() == firstName.ToLower())
-                    .ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, $"Error occured while finding the Customers named: {firstName}");
-                return null;
-            }
-        }
-
-        public async Task<List<Customer>> GetByMiddleName(string middleName)
-        {
-            try
-            {
-                _logger.LogInformation($"The Customers named: {middleName}, have been found.");
-                return await _context.Customers
-                    .Where(a => a.Middle_Name.ToLower() == middleName.ToLower())
-                    .ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, $"Error occured while finding the Customers named: {middleName}");
-                return null;
-            }
-        }
-
-
-        public async Task<List<Customer>> GetByLastName(string lastName)
-        {
-            try
-            {
-                _logger.LogInformation($"The Customers named: {lastName}, have been found.");
-                return await _context.Customers
-                    .Where(a => a.Last_Name.ToLower() == lastName.ToLower())
-                    .ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, $"Error occured while finding the Customers named: {lastName}");
-                return null;
-            }
-        }
-
-
-        public async Task<bool> Add(Customer customer)
-        {
-            try
-            {
-                _context.Customers.Add(customer);
-                await _context.SaveChangesAsync();
-                _logger.LogInformation($"Customer: {customer} added successfully.");
-                return true;
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, $"Error occurred while adding a Customer: {customer}");
-                return false;
-            }
-        }
-
-
-        public async Task<Customer> Edit(Guid guid, CustomerData customerData)
-        {
-            try
-            {
-                Customer? customer = await _context.Customers.FindAsync(guid);
-
-                if (customer == null)
+                if (ExistsByID(guid))
                 {
-                    _logger.LogWarning($"The Customer: {guid}, is not found");
-                    return null;
-                }
+                    Customer? customer = await _context.Customers.FindAsync(guid);
 
-                else
-                {
                     customer.TIN = customerData.TIN;
                     customer.First_Name = customerData.First_Name;
                     customer.Middle_Name = customerData.Middle_Name;
@@ -165,25 +202,85 @@ namespace CarDealer.Services
                     _context.Customers.Update(customer);
                     await _context.SaveChangesAsync();
 
-                    _logger.LogInformation($"The Customer: {guid}, has been updated successfully.");
+                    _logger.LogInformation($"The Customer {guid} has been updated successfully.");
                     return customer;
+                }
+                else
+                {
+                    _logger.LogWarning($"Can not update the Customer {guid}");
+                    return null;
                 }
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"Error occurred while updating the Customer: {guid}.");
+                _logger.LogError(ex, $"Error occurred while updating the Customer {guid}.");
                 return null;
             }
         }
+        #endregion
 
 
-        public async Task<bool> Remove(Customer customer)
+        #region Delete By Customer ID
+        public async Task<string> DeleteByID(Guid guid)
         {
-            _logger.LogInformation($"{customer.Customer_ID} is not more available.");
-            _context.Customers.Remove(customer);
+            try
+            {
+                if (ExistsByID(guid))
+                {
+                    Customer? customer = await _context.Customers.FindAsync(guid);
+                    _context.Customers.Remove(customer);
+                    _context.SaveChanges();
+                    _logger.LogInformation($"The Customer {customer} has been deleted successfully.");
 
-            await _context.SaveChangesAsync();
-            return true;
+                    return $"The Customer {customer} has been deleted successfully.";
+                }
+                else
+                {
+                    _logger.LogWarning($"Can not delete the Customer {guid}");
+
+                    return $"Can not delete the Customer {guid}";
+                }
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"Error occurred while deleting the Customer {guid}.");
+                return $"Error occurred while deleting the Customer {guid}.";
+            }
         }
+        #endregion
+
+
+        #region Exists By ID
+        public bool ExistsByID(Guid guid)
+        {
+            if (_context.Customers.Any(c => c.Customer_ID == guid))
+            {
+                _logger.LogInformation("The Customer Exists!");
+                return true;
+            }
+            else
+            {
+                _logger.LogInformation("The Customer Does Not Exists.");
+                return false;
+            }
+        }
+        #endregion
+
+
+        #region Exists TIN
+        public bool ExistsTIN(string tin)
+        {
+            if (_context.Customers.Any(c => c.TIN == tin))
+            {
+                _logger.LogInformation("The Customer Exists.");
+                return true;
+            }
+            else
+            {
+                _logger.LogInformation("The Customer Does Not Exists.");
+                return false;
+            }
+        }
+        #endregion
     }
 }
